@@ -1,16 +1,16 @@
 name: speech_to_text
-description: A Flutter plugin that exposes device specific speech to text recognition capability. 
-version: 6.1.1
+description: A Flutter plugin that exposes device specific speech to text recognition capability.
+version: 6.4.1
 homepage: https://github.com/csdcorp/speech_to_text
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.10.0"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  speech_to_text_platform_interface: ^2.0.1
+  speech_to_text_platform_interface: ^2.1.0
   speech_to_text_macos: ^1.0.2
   json_annotation: ^4.0.0
   clock: ^1.0.1
@@ -23,11 +23,12 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^1.0.0
-  json_serializable: ^4.0.2
-  fake_async: ^1.0.1
-  mockito: ^5.0.0
-  plugin_platform_interface: ^2.0.0
+  build_runner: ^2.4.4
+  json_serializable: ^6.7.0
+  fake_async: ^1.3.1
+  mockito: ^5.4.1
+  plugin_platform_interface: ^2.1.4
+  flutter_lints: ^3.0.0
 
 flutter:
   plugin:
