@@ -5,19 +5,19 @@ issue_tracker: https://github.com/csdcorp/speech_to_text/issues
 version: 1.0.2
 
 environment:
-  sdk: ">=2.17.5 <3.0.0"
-  flutter: ">=2.5.0"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  plugin_platform_interface: ^2.0.2
+  plugin_platform_interface: ^2.1.4
   speech_to_text_platform_interface: ^2.0.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.0
+  flutter_lints: ^2.0.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -25,10 +25,8 @@ dev_dependencies:
 # The following section is specific to Flutter packages.
 flutter:
   plugin:
-      implements: speech_to_text
-      platforms:
-        macos:
-          pluginClass: SpeechToTextMacosPlugin
-          dartPluginClass: SpeechToTextMacOS
-
-  
+    implements: speech_to_text
+    platforms:
+      macos:
+        pluginClass: SpeechToTextMacosPlugin
+        dartPluginClass: SpeechToTextMacOS
