@@ -3,19 +3,20 @@ description: A common platform interface for the speech_to_text plugin.
 homepage: https://github.com/csdcorp/speech_to_text/speech_to_text_platform_interface
 # NOTE: We strongly prefer non-breaking changes, even at the expense of a
 # less-clean API. See https://flutter.dev/go/platform-interface-breaking-changes
-version: 2.0.1
+version: 2.1.0
 
 dependencies:
   flutter:
     sdk: flutter
   meta: ^1.3.0
-  plugin_platform_interface: ^2.0.0
+  plugin_platform_interface: ^2.1.4
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mockito: ^5.0.7
-  pedantic: ^1.10.0
+  mockito: ^5.4.1
+  pedantic: ^1.11.1
+  flutter_lints: ^2.0.1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
