@@ -33,7 +33,7 @@ class SpeechToTextMacOS extends SpeechToTextPlatform {
   ///
   /// [debugLogging] controls whether there is detailed logging from the underlying
   /// plugins. It is off by default, usually only useful for troubleshooting issues
-  /// with a paritcular OS version or device, fairly verbose
+  /// with a particular OS version or device, fairly verbose
   @override
   Future<bool> initialize(
       {debugLogging = false, List<SpeechConfigOption>? options}) async {
