@@ -81,7 +81,7 @@ abstract class SpeechToTextPlatform extends PlatformInterface {
   ///
   /// [debugLogging] controls whether there is detailed logging from the underlying
   /// plugins. It is off by default, usually only useful for troubleshooting issues
-  /// with a paritcular OS version or device, fairly verbose
+  /// with a particular OS version or device, fairly verbose
   Future<bool> initialize(
       {debugLogging = false, List<SpeechConfigOption>? options}) {
     throw UnimplementedError('initialize() has not been implemented.');
